✨ (client/src): add new onboarding components and hooks for user setup ♻️ (client/src/pages/Dashboard.tsx): refactor to use new onboarding hook and improve navigation logic 🐛 (client/src/pages/Professionals.tsx): remove console.log for cleaner production code
